Bring Zookeeper Provider down into the pane to eliminate race
